Sniffer: 1. 解决了授时申请数据包中的检查下发字段不能正确显示的问题。 2. 添加了自组网数据包解析的功能。 3. 将导出Excel表格的字体设为了Courier New。

+++ YyWsnDeviceLibrary/Sensor.cs
@@ -11,7 +11,10 @@ namespace YyWsnDeviceLibrary
 
         public int PSensorSN { get; set; }
 
-        public int Serial { get; set; }
+        /// <summary>
+        /// 序列号
+        /// </summary>
+        public UInt16 Serial { get; set; }
 
         public int CSensorSN { get; set; }
